:recycle: Move ExceptionData
- ExceptionData를 Chart에서 불러오던 것 ExceptionStatistics 에서 불러오도록 수정
